Skip steps if completed

+++ scripts/ubuntu_bootstrap.sh
@@ -20,15 +20,15 @@ sudo apt-get install -y apt-transport-https ca-certificates software-properties-
 sudo apt-get install -y curl firefox git neovim silversearcher-ag tree vim
 
 # install binaries
-! [[ -s /usr/local/bin/terraform ]] && curl -o /tmp/terraform.zip https://releases.hashicorp.com/terraform/0.9.6/terraform_0.9.6_linux_amd64.zip && \
+! [[ -e /usr/local/bin/terraform ]] && curl -o /tmp/terraform.zip https://releases.hashicorp.com/terraform/0.9.6/terraform_0.9.6_linux_amd64.zip && \
   unzip /tmp/terraform.zip && \
   sudo mv terraform /usr/local/bin
-! [[ -s /usr/local/bin/packer ]] && curl -o /tmp/packer.zip https://releases.hashicorp.com/packer/1.0.0/packer_1.0.0_linux_amd64.zip && \
+! [[ -e /usr/local/bin/packer ]] && curl -o /tmp/packer.zip https://releases.hashicorp.com/packer/1.0.0/packer_1.0.0_linux_amd64.zip && \
   unzip /tmp/packer.zip && \
   sudo mv packer /usr/local/bin
 
 # install docker
-if ! [[ -s /usr/bin/docker ]]; then
+if ! [[ -e /usr/bin/docker ]]; then
   curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
   sudo add-apt-repository \
     "deb [arch=amd64] https://download.docker.com/linux/ubuntu yakkety stable \
@@ -40,15 +40,15 @@ if ! [[ -s /usr/bin/docker ]]; then
 fi
 
 # install toolchains
-rvm || \curl -sSL https://get.rvm.io | bash -s stable --ruby --gems=bundler,rails,nokogiri
-go || sudo apt-get install -y golang-go
-cargo || curl -sSf https://static.rust-lang.org/rustup.sh | sh
+! [[ -e ~/.rvm ]] && \curl -sSL https://get.rvm.io | bash -s stable --ruby --gems=bundler,rails,nokogiri
+! [[ -e /usr/bin/go ]] && sudo apt-get install -y golang-go
+! [[ -e /usr/local/bin/cargo ]] && curl -sSf https://static.rust-lang.org/rustup.sh | sh
 
 # install heroku tooling
-heroku --version || curl https://toolbelt.heroku.com/install-ubuntu.sh | sh && git checkout .bashrc
+! [[ -e /usr/local/heroku ]] && curl https://toolbelt.heroku.com/install-ubuntu.sh | sh && git checkout .bashrc
 
 # configure dotfiles
-if ! [ -d .git ]; then
+if ! [ -e ~/.git ]; then
   git init .
   git remote add -t \* -f origin https://github.com/charlieegan3/dotfiles.git
   git fetch origin
@@ -56,14 +56,19 @@ if ! [ -d .git ]; then
 fi
 
 # configure neovim
-curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs \
-        https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-nvim +PlugInstall +qall
-ln -sf .config/nvim/init.vim .vim_config
+if ! [ -e ~/.config/nvim/autoload/plug.vim ]; then
+  curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs \
+          https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+  nvim +PlugInstall +qall
+  ln -sf .config/nvim/init.vim .vim_config
+fi
 
 # gnome settings
 gsettings set org.gnome.desktop.background show-desktop-icons true
 gsettings set org.gnome.desktop.peripherals.touchpad tap-to-click true
 gsettings set org.gnome.desktop.sound event-sounds false
-# file new profile required?
-curl https://raw.githubusercontent.com/chriskempson/base16-gnome-terminal/master/base16-londontube.light.sh | bash
+if ! [ -e ~/resources/theme-installed ]; then
+  # new terminal profile required
+  bash ~/resources/base16-londontube.light.sh
+  touch ~/resources/theme-installed
+fi
